Successfully added hours to db

+++ db/place.js
@@ -20,7 +20,7 @@ module.exports = function(sequelize, Sequelize) {
       type: Sequelize.DOUBLE
     },
     hours: {
-      type: Sequelize.ARRAY(Sequelize.ARRAY(Sequelize.TINYINT))
+      type: Sequelize.ARRAY(Sequelize.ARRAY(Sequelize.INTEGER))
     }
   }, {
     timestamps: false,
